+++ ext/FlexiChainsMakieExt/forestplot.jl
@@ -18,9 +18,9 @@ function _draw_point_and_intervals!(ax, y, color, point_val, interval_sets)
     return Makie.scatter!(ax, [point_val], [y]; color=color, markersize=10)
 end
 
-###############
+####################
 # Makie.forestplot #
-###############
+####################
 
 """
     FlexiChains.Makie.forestplot(
@@ -71,9 +71,10 @@ function FC.Makie.forestplot(
     legend=(;),
     kwargs...,
 )
-    chn = FC.PlotUtils.subset_and_split_chain(chn, param_or_params)
+    chn, plot_names = FC.PlotUtils.subset_and_split_chain(chn, param_or_params)
     keys_to_plot = keys(chn)
     isempty(keys_to_plot) && throw(ArgumentError("no parameters to plot"))
+    kstrs = [FC.PlotUtils.get_plot_param_name(k, plot_names) for k in keys_to_plot]
     nparams = length(keys_to_plot)
     fig = Makie.Figure(;
         size=(
@@ -89,6 +90,7 @@ function FC.Makie.forestplot(
         FC.PlotUtils.FlexiChainForest(
             chn,
             collect(keys_to_plot),
+            kstrs,
             pool_chains,
             point,
             interval,
@@ -118,13 +120,15 @@ function FC.Makie.forestplot(
     axis=(;),
     kwargs...,
 )
-    chn = FC.PlotUtils.subset_and_split_chain(chn, param_or_params)
+    chn, plot_names = FC.PlotUtils.subset_and_split_chain(chn, param_or_params)
     ks = collect(keys(chn))
+    kstrs = [FC.PlotUtils.get_plot_param_name(k, plot_names) for k in ks]
     return FC.Makie.forestplot!(
         Makie.Axis(grid; _default_forestplot_axis()..., axis...),
         FC.PlotUtils.FlexiChainForest(
             chn,
             ks,
+            kstrs,
             pool_chains,
             point,
             interval,
@@ -146,13 +150,15 @@ function FC.Makie.forestplot!(
     pool_chains::Bool=false,
     kwargs...,
 )
-    chn = FC.PlotUtils.subset_and_split_chain(chn, param_or_params)
+    chn, plot_names = FC.PlotUtils.subset_and_split_chain(chn, param_or_params)
     ks = collect(keys(chn))
+    kstrs = [FC.PlotUtils.get_plot_param_name(k, plot_names) for k in ks]
     return FC.Makie.forestplot!(
         ax,
         FC.PlotUtils.FlexiChainForest(
             chn,
             ks,
+            kstrs,
             pool_chains,
             point,
             interval,
@@ -233,7 +239,7 @@ function FC.Makie.forestplot!(ax::Makie.Axis, d::FC.PlotUtils.FlexiChainForest; 
             end
         end
     end
-    ax.yticks = (Float64.(nparams:-1:1), map(k -> string(k.name), params))
+    ax.yticks = (Float64.(nparams:-1:1), d.param_names)
     Makie.ylims!(ax, 0.5, nparams + 0.5)
     return Makie.AxisPlot(ax, p)
 end