+++ ext/FlexiChainsMakieExt/hist.jl
@@ -1,5 +1,5 @@
-function _default_histogram_axis(k::FC.ParameterOrExtra)
-    return (xlabel="value", ylabel="probability", title=string(k.name))
+function _default_histogram_axis()
+    return (xlabel="value", ylabel="probability")
 end
 
 HIST_DOCSTRING = """
@@ -53,7 +53,7 @@ for f in (:hist, :stephist)
             legend=(;),
             kwargs...,
         )
-            chn = FC.PlotUtils.subset_and_split_chain(chn, param_or_params)
+            chn, plot_names = FC.PlotUtils.subset_and_split_chain(chn, param_or_params)
             keys_to_plot = keys(chn)
             isempty(keys_to_plot) && throw(ArgumentError("no parameters to plot"))
             nrows, ncols, figure =
@@ -62,8 +62,14 @@ for f in (:hist, :stephist)
             # This order means that plots go from left to right before going to the next row
             indices = Iterators.product(1:ncols, 1:nrows)
             for ((col, row), k) in zip(indices, keys_to_plot)
+                kstr = FC.PlotUtils.get_plot_param_name(k, plot_names)
                 a, p = Makie.$f!(
-                    Makie.Axis(figure[row, col]; _default_histogram_axis(k)..., axis...),
+                    Makie.Axis(
+                        figure[row, col];
+                        _default_histogram_axis()...,
+                        title=kstr,
+                        axis...,
+                    ),
                     FC.PlotUtils.FlexiChainHistogram(chn, k, pool_chains);
                     kwargs...,
                 )
@@ -82,10 +88,11 @@ for f in (:hist, :stephist)
         function Makie.$f(grid::MakieGrids, chn::FC.FlexiChain, param; axis=(;), kwargs...)
             # TODO: Error if there is already something at the grid position?
             # See e.g. https://github.com/rafaqz/DimensionalData.jl/blob/6db30de4b2e1fc7f8611b7e1dc3f89dc02c78598/ext/DimensionalDataMakieExt.jl#L85-L96
-            chn = FC.PlotUtils.subset_and_split_chain(chn, param)
+            chn, plot_names = FC.PlotUtils.subset_and_split_chain(chn, param)
             k = only(keys(chn))
+            kstr = FC.PlotUtils.get_plot_param_name(k, plot_names)
             return Makie.$f!(
-                Makie.Axis(grid; _default_histogram_axis(k)..., axis...),
+                Makie.Axis(grid; _default_histogram_axis()..., title=kstr, axis...),
                 chn,
                 param;
                 kwargs...,
@@ -98,7 +105,7 @@ for f in (:hist, :stephist)
             pool_chains::Bool=false,
             kwargs...,
         )
-            chn = FC.PlotUtils.subset_and_split_chain(chn, param)
+            chn, _ = FC.PlotUtils.subset_and_split_chain(chn, param)
             k = only(keys(chn))
             a, p = Makie.$f!(
                 ax,