+++ ext/FlexiChainsMakieExt/meanplot.jl
@@ -1,5 +1,5 @@
-function _default_meanplot_axis(k::FC.ParameterOrExtra)
-    return (xlabel="iteration number", ylabel="mean", title=string(k.name))
+function _default_meanplot_axis()
+    return (xlabel="iteration number", ylabel="mean")
 end
 
 """
@@ -26,15 +26,16 @@ function FC.Makie.meanplot(
     legend=(;),
     kwargs...,
 )
-    chn = FC.PlotUtils.subset_and_split_chain(chn, param_or_params)
+    chn, plot_names = FC.PlotUtils.subset_and_split_chain(chn, param_or_params)
     keys_to_plot = keys(chn)
     isempty(keys_to_plot) && throw(ArgumentError("no parameters to plot"))
     nrows, ncols, figure = setup_figure_and_layout(length(keys_to_plot), 1, layout, figure)
     a, p = nothing, nothing
     indices = Iterators.product(1:ncols, 1:nrows)
     for ((col, row), k) in zip(indices, keys_to_plot)
+        kstr = FC.PlotUtils.get_plot_param_name(k, plot_names)
         a, p = FC.Makie.meanplot!(
-            Makie.Axis(figure[row, col]; _default_meanplot_axis(k)..., axis...),
+            Makie.Axis(figure[row, col]; _default_meanplot_axis()..., title=kstr, axis...),
             FC.PlotUtils.FlexiChainMean(chn, k);
             kwargs...,
         )
@@ -48,17 +49,18 @@ end
 # Single axis plotting #
 ########################
 function FC.Makie.meanplot(grid::MakieGrids, chn::FC.FlexiChain, param; axis=(;), kwargs...)
-    chn = FC.PlotUtils.subset_and_split_chain(chn, param)
+    chn, plot_names = FC.PlotUtils.subset_and_split_chain(chn, param)
     k = only(keys(chn))
+    kstr = FC.PlotUtils.get_plot_param_name(k, plot_names)
     return FC.Makie.meanplot!(
-        Makie.Axis(grid; _default_meanplot_axis(k)..., axis...),
+        Makie.Axis(grid; _default_meanplot_axis()..., title=kstr, axis...),
         chn,
         param;
         kwargs...,
     )
 end
 function FC.Makie.meanplot!(ax::Makie.Axis, chn::FC.FlexiChain, param; kwargs...)
-    chn = FC.PlotUtils.subset_and_split_chain(chn, param)
+    chn, plot_names = FC.PlotUtils.subset_and_split_chain(chn, param)
     k = only(keys(chn))
     a, p = FC.Makie.meanplot!(ax, FC.PlotUtils.FlexiChainMean(chn, k); kwargs...)
     return Makie.AxisPlot(a, p)